+++ android-sdk-ui/assets/appboy-html-in-app-message-javascript-component.js
@@ -43,3 +43,4 @@ var appboyBridge = {
     return this.appboyBridgeUserObject;
   }
 };
+window.dispatchEvent(new Event("ab.BridgeReady"));